+++ modules/genealogy-people-filament/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Added the person attributes editor to the Filament resource.
+
 ## 1.0.0
 
 - Initial presentation boundary.
-